Track changes on apps, pagination and fixes
Added history of changes for given app. User can check who, when and what changes was made on the app including on the runtime configuration.
Added pagination for history and notifications of an app. Every page will show 25 results.
Cosmetics on notification list, so it looks pretty with long urls and message bodies.
Fixed toggles button on creating a new app dialog.
Fixed input fields validation for firebase configuration showing that configuration is already uploaded.